requirements.txt
bme680==2.0.0
certifi==2024.7.4
influxdb-client==1.44.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
reactivex==4.0.4
setuptools==72.1.0
six==1.16.0
smbus2==0.4.3
typing_extensions==4.12.2
urllib3==2.2.2
